+++ .eslintrc
@@ -18,6 +18,8 @@
     "class-methods-use-this": 0,
     "linebreak-style": 0,
     "max-len": [1,600,2],
-    "no-plusplus": [2, { "allowForLoopAfterthoughts": true }]
+    "no-plusplus": [2, { "allowForLoopAfterthoughts": true,
+    "import/no-extraneous-dependencies": ["error", {"devDependencies": false, "optionalDependencies": false, "peerDependencies": false}]
+    }]
   }
 }
