+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "error clean",
+  "name": "errorClean",
   "version": "1.0.0",
   "description": "",
   "main": "public/main.js",
@@ -33,6 +33,7 @@
     "bootstrap": "^5.1.0",
     "dotenv": "^10.0.0",
     "firebase": "^8.2.0",
-    "sass": "^1.51.0"
+    "sass": "^1.51.0",
+    "uuid": "^11.0.5"
   }
 }
