+++ utils/sample_data/voldy.js
@@ -1,0 +1,3 @@
+const voldysArmy = [];
+
+export default voldysArmy;
